requirements.txt
# ========= CORE =========
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
numpy>=1.26.0
tqdm>=4.66.1

# ========= LLM & OPENAI =========
openai>=1.45.0

# ========= DATABASE / SUPABASE =========
supabase>=2.3.3
postgrest>=0.14.3

# ========= USD / 3D GEOMETRY =========
# Pixar USD Python bindings (provided by pip install pxr or conda package)
usd-core>=24.5.0
shapely>=2.0.3
trimesh>=4.3.2
scipy>=1.14.0

# ========= MACHINE LEARNING (LayoutVLM) =========
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.42.0
einops>=0.8.0

# ========= VISUALIZATION / DEBUGGING =========
matplotlib>=3.9.0
Pillow>=10.3.0
opencv-python>=4.9.0.80

# ========= API SERVER =========
fastapi>=0.115.0
uvicorn[standard]>=0.30.1

# ========= UTILITIES =========
pandas>=2.2.2
aiofiles>=23.2.1

# ========= OPTIONAL DEV / DEBUG =========
black>=24.3.0
flake8>=7.1.0
